backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pyyaml
openai
python-dotenv
polars
python-multipart
redis
